modified "JwtGenerator" and add a claim for user's id.
modified "JwtValidator" and validate user id, email and check if user is available or not and if user is active or not.

modified "AuthUserUC" and made it's out put a tuple<User,bool> and its usage in controllers.

add a xml file for comments

modified is program.cs is huge :
add "AddAuthentication" and "AddJwtBearer" to just implement it and practive for coding while  have custom authentication service base on jwt.
so i also add a method to my JwtValidato named "Execute" and use in TokenValidated event on Jwt. will be commented soon.
add comments in swagger ui. also encapsulate swagger uUIs based on version of api. ps: project has two api version.