add LogOut method in "AuthenticationUC" to manage logout of user and use authorization to remvoe tokens from database after a successful logout.
add to methods in "IAuthorizationServices":
DeleteRangeTokensAsync to remove tokens of user who wants to log out.
CheckExistToken to limit the users and wont let them to authorize with their removed tokens.

add CheckExistToken section in JwtValidator.

add an action to AccountController dedicated to log out the user.

add an OpenApiSecurityScheme in program.cs to swagger to get access token (JWT) from user and authorize them in swagger.

Changed authentication approach from using a Middleware to authorize jwt to user AddAuthentication and AddJwtBearer.
ps: previous way was authorizing all coming requests and when clients send a request to an action that did not need Authorisation, application will returned an exception cause request does not include an access token.